Metashape - Extract ImageProcessing class from process-images script
Break out the class to allow for better documentation of the script
entry point and reduce the class file length for improved readability.
